Implemented assumptions about event behavior and added more tests

+++ src/main/kotlin/parkinglot/simulator/domain/repository/VehicleTransitRepository.kt
@@ -3,5 +3,5 @@ package parkinglot.simulator.domain.repository
 interface VehicleTransitRepository {
     fun addVehicleInTransit(licensePlate: String)
     fun removeVehicleInTransit(licensePlate: String)
-    fun getNumberOfVehiclesInTransit(): Int
+    fun getNumberOfVehiclesInTransit(): Long
 }
